chore(scripts): fix scripts

+++ scripts/generate-gomock.sh
@@ -4,20 +4,38 @@ TAG="[generate-gomock.sh]"
 
 MOCK_PREFIX="mock"
 
-GO_FILES=$(ls ${PKG_DIR}/*.go | xargs -n 1 basename)
+USAGE="Usage: PKG_DIR={RELATIVE_PATH} $0"
 
+if [ "${PKG_DIR}" == "" ]; then
+    echo ${USAGE} >&2
+    exit 1
+fi
+
+WORKSPACE=$(echo ${PKG_DIR} | tr -d " ")
+
+GO_FILES=$(ls ${PKG_DIR}/*.go | xargs -n 1 basename)
 for GO_FILE in ${GO_FILES[@]}
 do
+
+    if [[ ${GO_FILE} == *"test.go" ]]; then
+        continue
+    fi
+
     PKG=$(basename ${PKG_DIR})
     MOCK_FILE=${PKG_DIR}/${MOCK_PREFIX}_${GO_FILE}
     mockgen -package ${PKG} -source ${PKG_DIR}/${GO_FILE} -destination ${MOCK_FILE}
 
-    # remove if the generated file doesn't include `struct` (useless mock files)
-    if cat ${MOCK_FILE} | grep -q --line-buffered struct;
+    if [ -f ${MOCK_FILE} ];
     then
-        echo -e "Generated: ${MOCK_FILE}"
-    else
-        rm ${MOCK_FILE}
+
+      # remove if the generated file doesn't include `struct` (useless mock files)
+      if cat ${MOCK_FILE} | grep -q --line-buffered struct;
+      then
+          echo -e "\t$TAG Generated: ${MOCK_FILE}"
+      else
+          echo -e "\t$TAG Removing: ${MOCK_FILE}"
+          rm ${MOCK_FILE}
+      fi
     fi
 
 done
